requirements.txt
Django==2.2.4
django-allauth==0.32.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
requests==2.21.0
django-filter==2.2.0
django-bootstrap-form
django-bootstrap3

